Revert "Add async functions support in router (#854)"
This reverts commit 9f17c9b0becd37eac425fac96d292012bf0e59e3.
