+++ lib/utils/router/src/index.js
@@ -1,10 +1,12 @@
 'use strict';
 
 // Small Express router that runs request handlers written as ES6 generators
-// using Node co. On top of that the router supports batching
+// using Node co. On top of that the router does a few useful things, including
+// some logging and error handling using a Node domain.
 
 const _ = require('underscore');
 const express = require('express');
+// const domain = require('domain');
 const url = require('url');
 const yieldable = require('abacus-yieldable');
 const transform = require('abacus-transform');
@@ -21,26 +23,33 @@ const allKeys = _.allKeys;
 const debug = require('abacus-debug')('abacus-router');
 const edebug = require('abacus-debug')('e-abacus-router');
 
-// https://medium.com/@Abazhenov/using-async-await-in-express-with-node-8-b8af872c0016
-const asyncMiddleware = (fn) => (req, res, next) =>
-  Promise.resolve(fn(req, res, next)).then((value) => next(undefined, value)).catch(next);
-
 // Convert a middleware function which can be a regular Express middleware or a
 // generator to a regular Express middleware function.
 const callbackify = (m, trusted) => {
-  const mfunc = yieldable.isAsyncFunction(m) ? asyncMiddleware(m) : yieldable.functioncb(m);
+  // If the callback is a regular function just use it as-is, otherwise it's
+  // a generator and we need to wrap it using the co module
+  const mfunc = yieldable.functioncb(m);
 
   // Return a middleware function. Middleware functions can be of the form
-  // function(req, res, next) or function(err, req, res, next)
+  // function(req, res, next) or function(err, req, res, next) so we need to
+  // support both forms
   return function() {
     const next = arguments[arguments.length - 1];
     const res = arguments[1];
     const params = toArray(arguments).slice(0, arguments.length - 1);
 
-    // Pass errors down the middleware stack
+    // Pass errors down the middleware stack, if the middleware is
+    // un-trusted then we mark the error with bailout flag to trigger our
+    // server bailout logic
     const error = (err, type) => {
       edebug('Route error - %s - %o', type, err);
       debug('Route error - %s - %o', type, err);
+      // TODO Re-enable after ensuring that route implementations catch all
+      // errors except really fatal errors
+      /*
+      if(!trusted && !err.status && !err.statusCode)
+        err.bailout = true;
+      */
       next(err);
     };
 
@@ -98,9 +107,45 @@ const route = (original, trusted) => {
   };
 };
 
+// Return an Express middleware that uses a Node domain to run the middleware
+// stack and handle any errors not caught in async callbacks
+const catchall = (trusted) => {
+  return (req, res, next) => {
+    // TODO Re-enable this after we understand why unrelated contexts are
+    // incorrectly captured by this domain
+    /*
+    const d = domain.create();
+    d.on('error', (err) => {
+      debug('Route domain error %o', err);
+
+      // Pass the error down the middleware stack, if the router runs
+      // un-trusted middleware then mark it with a bailout flag to
+      // trigger our server bailout logic
+      // Warning: mutating variable err
+      if(!trusted && !err.status && !err.statusCode)
+        err.bailout = true;
+      next(err);
+    });
+
+    // Because req and res were created before this domain existed,
+    // we need to explicitly add them.  See the explanation of implicit
+    // vs explicit binding in the Node domain docs.
+    d.add(req);
+    d.add(res);
+
+    // Run the middleware stack in our new domain
+    d.run(next);
+    */
+    next();
+  };
+};
+
 // Return an Express router middleware that works with generators
 const router = (trusted) => {
   const r = express.Router();
+
+  // Catch all errors down the middleware stack using a Node domain
+  r.use(catchall(trusted));
 
   // Monkey patch the router function with our implementation of the use
   // and route functions
