+++ lib/utils/yieldable/src/index.js
@@ -9,6 +9,7 @@ const co = require('co');
 
 const extend = _.extend;
 const isFunction = _.isFunction;
+const isObject = _.isObject;
 const object = _.object;
 const pairs = _.pairs;
 const map = _.map;
@@ -16,19 +17,11 @@ const functions = _.functions;
 const rest = _.rest;
 const bind = _.bind;
 
-const isPromise = (obj) => 'function' === typeof obj.then;
+// Return true if the given function is a generator
+const isGenerator = (fn) => isFunction(fn) && Object.getPrototypeOf(fn) !== Object.getPrototypeOf(Function);
 
-const isGenerator = (obj) => 'function' === typeof obj.next && 'function' === typeof obj.throw;
-
-const isGeneratorFunction = (obj) => {
-  const constructor = obj.constructor;
-  if (!constructor) return false;
-  if ('GeneratorFunction' === constructor.name || 'GeneratorFunction' === constructor.displayName) return true;
-  return isGenerator(constructor.prototype);
-};
-
-const AsyncFunction = (async() => {}).constructor;
-const isAsyncFunction = (obj) => obj instanceof AsyncFunction;
+// Return true if the given function is actually a promise
+const isPromise = (fn) => isObject(fn) && isFunction(fn.then);
 
 // Convert a value to a function that returns that value
 const thunk = (value) => (cb) => cb(undefined, value);
@@ -44,7 +37,7 @@ const nthunkify = (fn) =>
 // a thunkifed function (a function that returns a thunk).
 const generator = (fn) => {
   // Return a generator as is
-  if (isGeneratorFunction(fn)) return fn;
+  if (isGenerator(fn)) return fn;
 
   // Convert to a generator
   return extend(
@@ -63,7 +56,7 @@ const generator = (fn) => {
 // using co), and promises.
 const functioncb = (fn) => {
   // Convert an ES6 stream generator to a function with callback using co
-  if (isGeneratorFunction(fn))
+  if (isGenerator(fn))
     return extend(co(fn), {
       // Preserve the original function name
       fname: fn.fname || fn.name
@@ -107,7 +100,7 @@ const nbind = (o, k) =>
 // that will generate a stream containing the single result of that function.
 const yieldable = (fn) => {
   // Return a generator as is
-  if (isGeneratorFunction(fn)) return fn;
+  if (isGenerator(fn)) return fn;
 
   // Convert a promise
   if (isPromise(fn)) return generator(nthunkify(functioncb(fn)));
@@ -127,7 +120,3 @@ module.exports.yieldable = yieldable;
 module.exports.thunk = thunk;
 module.exports.functioncb = functioncb;
 module.exports.promise = promise;
-module.exports.isPromise = isPromise;
-module.exports.isGenerator = isGenerator;
-module.exports.isGeneratorFunction = isGeneratorFunction;
-module.exports.isAsyncFunction = isAsyncFunction;
