+++ lib/utils/yieldable/src/test/test.js
@@ -163,41 +163,4 @@ describe('abacus-yieldable', () => {
       done();
     });
   });
-
-  it('checks for promises', () => {
-    expect(yieldable.isPromise(new Promise((resolve) => {
-      resolve(2);
-    }))).to.equal(true);
-    expect(yieldable.isPromise(() => {
-      return 2;
-    })).to.equal(false);
-  });
-
-  const gen = function*() {
-    yield 1;
-  };
-
-  it('checks for generators', () => {
-    expect(yieldable.isGenerator(gen())).to.equal(true);
-    expect(yieldable.isGenerator({})).to.equal(false);
-  });
-
-  it('checks for generator functions', () => {
-    expect(yieldable.isGeneratorFunction(function*() {
-      return 2;
-    })).to.equal(true);
-    expect(yieldable.isGeneratorFunction(() => {
-      return 2;
-    })).to.equal(false);
-  });
-
-  it('checks for async functions', () => {
-    expect(yieldable.isAsyncFunction(async() => {
-      return 2;
-    })).to.equal(true);
-    expect(yieldable.isAsyncFunction(() => {
-      return 2;
-    })).to.equal(false);
-  });
-
 });
